codecrafters submit [skip ci]

+++ app/resp_parser.py
@@ -79,7 +79,7 @@ class LpopCommand:
 class BlpopCommand:
     """BLPOP is a blocking variant of the LPOP command. It waits for an element to become available on a list before popping it."""
     list_key: str
-    timeout: int
+    timeout: float
 
 
 @dataclass
@@ -301,7 +301,7 @@ def parse_command(data: bytes):
         elif cmd_name == 'BLPOP':
             if len(args) < 1 or len(args) > 2:
                 return CommandError("wrong number of arguments for 'blpop' command")
-            return BlpopCommand(list_key=str(args[0]), timeout=int(args[1]))
+            return BlpopCommand(list_key=str(args[0]), timeout=float(args[1]))
 
         else:
             return CommandError(f"unknown command '{cmd_name}'")
